data_pipelines/stocks/services/alphavantage_extractor/pyproject.toml
[project]
name = "alphavantage-extractor"
version = "0.1.0"
description = "Alpha Vantage data extraction service"
requires-python = ">=3.11"
dependencies = [
    "alphavantage",
    "google-cloud-storage>=3.4.0",
    "google-cloud-bigquery>=3.38.0",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
alphavantage = { workspace = true }

# Optional: Add scripts if you want CLI commands
[project.scripts]
alphavantage-extract = "alphavantage_extractor.main:main"